core: move pragma statement bytecode generator to its own file.
